TaxService

+++ Services/CustomerService/DTOs/ItemListDto.cs
@@ -6,6 +6,6 @@ namespace CustomerService.DTOs
         public int CartId { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
     }
 }
