+++ Services/InventoryService/DTOs/ProductInventoryDto.cs
@@ -8,7 +8,7 @@ namespace InventoryService.DTOs
         public string ProductName { get; set; }
         public string ProductType { get; set; }
         public string ProductSubType { get; set; }
-        public decimal UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
         public float Cost { get; set; }
         public int InStock { get; set; }
         public int VendorId { get; set; }
