+++ Services/OrderManagementService/Models/Order.cs
@@ -11,7 +11,7 @@ namespace OrderManagementService.Models
         public int OrderId { get; set; }
         public int CustomerId { get; set; }
         public DateTime OrderDate { get; set; }
-        public decimal TotalAmount { get; set; }
+        public float TotalAmount { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
     }
